Fix previous/next links for documentation
When moving to the motif-docs the sidebar config structure was changed
but the logic in sidebar.ts was not updated to handle the new structure.

This commit changes the logic to use the `href` values provided rather
than trying to make the slug from the title since the manually chosen
slug (i.e. the `href` value) is often different from a slugified title.

We also expand the logic to properly handle section overview pages (or
lack thereof) and make sure that we don't cause errors for people
manually manipulating the URL (e.g. `/docs/guides/get-started` ->
`/docs/guides`).

The commit cleans up sidebar functions that are not used in the codebase
to make maintenance easier for the next person.